modularize password inputs into separate components
