+++ src/components/Inputs/PasswordInputs/index.ts
@@ -1,0 +1,3 @@
+export * from './CurrentPassword'
+export * from './NewPassword'
+export * from './ConfirmNewPassword'
